refactor(mastracode): split workspace files panel into context, toggle and surface
The single WorkspaceFilesPanel module held the context, the provider, the
header toggle and the docked card, which meant chat components importing only
the inset class string pulled the whole viewer tree with it.

Layout constants now live in workspace-viewer/layout.ts, the context follows
the repo's Context/Provider/use split, and the toggle, surface and shared
content are one component per file.
