+++ mastracode/factory-ui/src/ui/domains/workspace-viewer/context/WorkspaceFilesContext.ts
@@ -1,0 +1,12 @@
+import { createContext } from 'react';
+
+export interface WorkspaceFilesApi {
+  open: boolean;
+  setOpen: (open: boolean) => void;
+  workspacePath?: string;
+  viewingFile: boolean;
+  setViewingFile: (viewing: boolean) => void;
+  canDock: boolean;
+}
+
+export const WorkspaceFilesContext = createContext<WorkspaceFilesApi | undefined>(undefined);
