+++ mastracode/factory-ui/src/ui/domains/workspace-viewer/index.ts
@@ -1,9 +1,3 @@
 export { renderedPaths } from './config';
 export type { RenderedWorkspacePath } from './config';
 export { WorkspaceViewerPanel } from './components/WorkspaceViewerPanel';
-export {
-  WorkspaceFilesPanelProvider,
-  WorkspaceFilesSurface,
-  WorkspaceFilesToggle,
-} from './components/WorkspaceFilesPanel';
-export { useThreadWorkspacePath } from './hooks/useThreadWorkspacePath';
